move all /components/App files to /components and update references
