+++ src/components/AppBar/CustomAppBar.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { AppBar, Button, Toolbar, Tooltip, Menu, useMediaQuery } from '@material-ui/core';
-import LoadSaveScheduleFunctionality from '../App/LoadSaveFunctionality';
+import LoadSaveScheduleFunctionality from '../LoadSaveFunctionality';
 import { Assignment } from '@material-ui/icons';
 import MenuIcon from '@material-ui/icons/Menu';
 import MenuItem from '@material-ui/core/MenuItem';
@@ -11,7 +11,7 @@ import { ReactComponent as Logo } from './logo.svg';
 import { ReactComponent as MobileLogo } from './mobile-logo.svg';
 import News from './News';
 import AboutPage from './AboutPage';
-import ConditionalWrapper from '../App/ConditionalWrapper';
+import ConditionalWrapper from '../ConditionalWrapper';
 import ImportStudyList from './ImportStudyList';
 
 const styles = {
